Move days rebase functions to RebaseDateTime
